+++ client/indexer-db/migrations/2024-10-07-133907_create_file/up.sql
@@ -2,7 +2,7 @@ CREATE TABLE file (
     id BIGSERIAL PRIMARY KEY,
     account BYTEA NOT NULL,
     file_key BYTEA NOT NULL,
-    bucket_id INTEGER NOT NULL,
+    bucket_id BIGINT NOT NULL,
     location BYTEA NOT NULL,
     fingerprint BYTEA NOT NULL,
     size BIGINT NOT NULL,
@@ -19,7 +19,7 @@ CREATE INDEX idx_file_bucket_id ON file(bucket_id);
 
 -- Create Bsp_File table
 CREATE TABLE bsp_file (
-    bsp_id INTEGER NOT NULL,
+    bsp_id BIGINT NOT NULL,
     file_id BIGINT NOT NULL,
     PRIMARY KEY (bsp_id, file_id),
     FOREIGN KEY (file_id) REFERENCES file(id) ON DELETE CASCADE
